% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{annotate_blast}
\alias{annotate_blast}
\title{Annotate blast results}
\usage{
annotate_blast(blast_out, metadata, dir = ".")
}
\arguments{
\item{blast_out}{The filename of the blast results}

\item{metadata}{The subset of the metadata obtained with the
\code{merge_files} or the \code{download_files} functions.}

\item{dir}{The directory where the fasta files are located.}
}
\value{
A \code{data.frame} with the annotated results
}
\description{
This function will add metadata to the blast results.
}
\details{
The blast results must contain an header with the following columns:

\enumerate{
  \item qseqid
  \item sseqid
  \item evalue
  \item qlen
  \item qlen
  \item pident
  \item qcovs
  \item length
  \item sstart
  \item send
}
}
\examples{
\dontrun{
  metadata <- get_demo_metadata()
  annotate_blast("test.out", metadata, dir = ".")
}

}
