% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_files.R
\name{download_files}
\alias{download_files}
\title{Download fasta files}
\usage{
download_files(n, metadata = NULL, dir = ".", merge = NULL,
  strict = TRUE, force = FALSE, verbose = FALSE, check_md5 = FALSE)
}
\arguments{
\item{n}{The names of the bacteria to download, correcponds to the
\code{organism_name} column from the \code{metadata} parameter.}

\item{metadata}{The metadata file downloaded with the
\code{fetch_bacteria_metadata} function. If \code{NULL}, the pre-computed
metadatas will be used. Default: \code{NULL}}

\item{dir}{The output directory. Default: \code{.}.}

\item{strict}{If \code{TRUE}, the names must be identical to the
\code{organism_name} column of the metadata object (by default, it's the
\code{bacteria} object). Otherwise, partial match will be allowed and the
matchning will not be case sensitive. Default: \code{TRUE}.}

\item{force}{If \code{TRUE}, the file will be downloaded even if it is
already present in the \code{dir/fna} directory. Default: \code{FALSE}.}

\item{verbose}{Print details of the downloading process. Default:
\code{FALSE}.}

\item{check_md5}{Check the md5 value of the files that are already
downloaded. Default: \code{FALSE}.}
}
\value{
Invisibly return the subset of metadata that was selected to
download.
}
\description{
Download the \code{_cds_from_genomic.fna.gz} files for a list of bacteria
}
\details{
This script will download all the \code{_cds_from_genomic.fna.gz} based on a
list of bacteria names.

The results will be saved in the `code{dir/fna/} directory.

To avoid redownloading the same files over and over, this function will
check by default if the file is already present with the correct md5sum
before downloading.
}
\examples{
\dontrun{
  download_files("Campylobacter_avium")
}

}
