% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{extract_fasta}
\alias{extract_fasta}
\title{Extract fasta sequences}
\usage{
extract_fasta(blast_out, metadata, merged_fasta)
}
\arguments{
\item{blast_out}{The filename of the blast results}

\item{metadata}{The subset of the metadata obtained with the
\code{merge_files} or the \code{download_files} functions.}

\item{merged_fasta}{The fasta obtained with the \code{merge_files} function}
}
\value{
TODO
}
\description{
Extract the sequences obtained after a blast analysis.
}
\details{
The blast results must contain an header with the following columns:

\enumerate{
  \item qseqid
  \item sseqid
  \item evalue
  \item qlen
  \item qlen
  \item pident
  \item qcovs
  \item length
  \item sstart
  \item send
}
}
\examples{
\dontrun{
  metadata <- get_demo_metadata()
  extract_fasta("test.out", metadata, "merged.fna")
}

}
