% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{fetch_bacteria_metadata}
\alias{fetch_bacteria_metadata}
\title{Genbank's bacterias metadata}
\usage{
fetch_bacteria_metadata()
}
\value{
A \code{data.frame} with the following columns:
\itemize{
  \item assembly_accession
  \item asm_name
  \item refseq_category
  \item organism_name
  \item infraspecific_name
  \item assembly_level
  \item version_status
  \item ftp_path
}
}
\description{
Fetch the metadata for all the bacterias in Genbank
}
\details{
Data is obtained from the \code{assembly_summary.txt} file:
ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/bacteria/assembly_summary.txt

Only the assembly at the Complete Genome level are kept.
}
\examples{
\dontrun{
  bacteria <- fetch_bacteria_metadata()
}

}
