% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Merge multiple assemblies}
\usage{
merge_files(n, output, dir = ".", metadata = NULL, strict = FALSE,
  force = FALSE)
}
\arguments{
\item{n}{The names of the bacteria to download, correcponds to the
\code{organism_name} column from the \code{metadata} parameter.}

\item{output}{The filename to save the merged fasta.}

\item{dir}{The download directory. Default: \code{.}.}

\item{metadata}{The metadata file downloaded with the
\code{fetch_bacteria_metadata} function. If \code{NULL}, the pre-computed
metadatas will be used. Default: \code{NULL}}

\item{strict}{If \code{TRUE}, the names must be identical to the
\code{organism_name} column of the metadata object (by default, it's the
\code{bacteria} object). Otherwise, partial match will be allowed and the
matchning will not be case sensitive. Default: \code{TRUE}.}

\item{force}{If \code{TRUE}, remove file if it already exists. Default:
\code{FALSE}.}
}
\value{
Invisibly return the subset of metadata that was selected to
download.
}
\description{
Merge multiple assemblies
}
\examples{
\dontrun{
  merge("Campylobacter avium", output = "c_avium.fna")
}

}
