% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacteria.R
\name{fetch_bacteria_metadata}
\alias{fetch_bacteria_metadata}
\title{Genbank's bacterias metadata}
\usage{
fetch_bacteria_metadata()
}
\value{
A \code{data.frame} with the following columns:
\itemize{
  \item First item
  \item Second item
}
}
\description{
Fetch the metadata for all the bacterias in Genbank
}
\details{
Data is obtained from the \code{assembly_summary.txt} file:
ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/bacteria/assembly_summary.txt
}
\examples{
\dontrun{
  bacteria <- fetch_bacteria_metadata()
}

}
